Fixed taskUpdate endpoint and put services in proper location
